// resources/sap/fe/core/converters/ManifestSettings.ts
export enum ActionType {
	DataFieldForAction = "ForAction",
	DataFieldForIntentBasedNavigation = "ForNavigation",
	Default = "Default",
	Primary = "Primary",
	Secondary = "Secondary",
	SwitchToActiveObject = "SwitchToActiveObject",
	SwitchToDraftObject = "SwitchToDraftObject",
	DefaultApply = "DefaultApply",
	Menu = "Menu",
	ShowFormDetails = "ShowFormDetails"
}
// ...
export type FormatOptionsType = {
	hasDraftIndicator?: boolean;
	semantickeys?: string[];
	textLinesEdit?: number;
	textMaxCharactersDisplay?: number;
	textExpandBehaviorDisplay?: string;
};
// ...
type ManifestFormElement = Positionable & {
	template: string;
	label?: string;
	formatOptions?: FormatOptionsType;
};

export type FormManifestConfiguration = {
	fields: ConfigurableRecord<ManifestFormElement>;
};